rename Timer component directory to smaller case timer
