+++ src/components/timer/EmptyState.tsx
