+++ src/components/timer/TimerBackground.tsx
