+++ src/components/timer/TimerControls.tsx
