+++ src/components/timer/index.ts
